+++ poseidon-server/src/main/java/net/minecraft/server/BlockSlowSand.java
@@ -9,7 +9,7 @@ public class BlockSlowSand extends Block {
     public AxisAlignedBB e(World world, int i, int j, int k) {
         float f = 0.125F;
 
-        return AxisAlignedBB.b((double) i, (double) j, (double) k, (double) (i + 1), (double) ((float) (j + 1) - f), (double) (k + 1));
+        return AxisAlignedBB.b(i, j, k, i + 1, (float) (j + 1) - f, k + 1);
     }
 
     public void a(World world, int i, int j, int k, Entity entity) {
