+++ poseidon-server/src/main/java/net/minecraft/server/ItemBlock.java
@@ -107,7 +107,7 @@ public class ItemBlock extends Item {
 
                 Block.byId[this.id].postPlace(world, i, j, k, l);
                 Block.byId[this.id].postPlace(world, i, j, k, entityhuman);
-                world.makeSound((double) ((float) i + 0.5F), (double) ((float) j + 0.5F), (double) ((float) k + 0.5F), block.stepSound.getName(), (block.stepSound.getVolume1() + 1.0F) / 2.0F, block.stepSound.getVolume2() * 0.8F);
+                world.makeSound((float) i + 0.5F, (float) j + 0.5F, (float) k + 0.5F, block.stepSound.getName(), (block.stepSound.getVolume1() + 1.0F) / 2.0F, block.stepSound.getVolume2() * 0.8F);
                 --itemstack.count;
             }
