+++ poseidon-server/src/main/java/net/minecraft/server/NoiseGeneratorOctaves.java
@@ -48,6 +48,6 @@ public class NoiseGeneratorOctaves extends NoiseGenerator {
     }
 
     public double[] a(double[] adouble, int i, int j, int k, int l, double d0, double d1, double d2) {
-        return this.a(adouble, (double) i, 10.0D, (double) j, k, 1, l, d0, 1.0D, d1);
+        return this.a(adouble, i, 10.0D, j, k, 1, l, d0, 1.0D, d1);
     }
 }
